Upgrade RepoStatus.ID from int to int64
This is needed on 32 bits platforms. Right now it is impossible to call
CreateStatus() on 32 bits executable because the ID returned by Github
overflows.

This is an API breaking change.

Fixes #806

+++ github/github-accessors.go
@@ -8565,7 +8565,7 @@ func (r *RepoStatus) GetDescription() string {
 }
 
 // GetID returns the ID field if it's non-nil, zero value otherwise.
-func (r *RepoStatus) GetID() int {
+func (r *RepoStatus) GetID() int64 {
 	if r == nil || r.ID == nil {
 		return 0
 	}
